# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API El Buen Sabor",
    "version": "1.0.0",
    "description": "API para gestión de pedidos de restaurante. Documentación automática.",
    "contact": {
      "name": "Dante Patroni",
      "url": "https://github.com/Dante-Patroni/el-buen-sabor-backend.git"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/platos": {
      "get": {
        "summary": "Obtiene el menú completo",
        "tags": [
          "Platos"
        ],
        "responses": {
          "200": {
            "description": "Lista de platos"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo plato",
        "tags": [
          "Platos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "precio": {
                    "type": "number"
                  },
                  "rubroId": {
                    "type": "integer"
                  },
                  "esMenuDelDia": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plato creado exitosamente"
          }
        }
      }
    },
    "/api/platos/{id}": {
      "put": {
        "summary": "Editar un plato (Precio, Menú del día)",
        "tags": [
          "Platos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "precio": {
                    "type": "number"
                  },
                  "esMenuDelDia": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plato actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un plato por ID",
        "tags": [
          "Platos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del plato a eliminar",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plato eliminado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string",
                      "example": "Plato eliminado correctamente"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plato no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/platos/{id}/imagen": {
      "post": {
        "summary": "Sube una foto para un plato",
        "tags": [
          "Platos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imagen": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen subida correctamente"
          }
        }
      }
    },
...
  },
...
}